+++ rcljava/include/org_ros2_rcljava_node_service_Client.h
@@ -1,4 +1,5 @@
 // Copyright 2016 Esteve Fernandez <esteve@apache.org>
+// Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
