+++ rcljava/src/main/java/org/ros2/rcljava/NativeUtils.java
@@ -1,4 +1,5 @@
-/* Copyright 2016 Open Source Robotics Foundation, Inc.
+/* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,8 +18,7 @@ package org.ros2.rcljava;
 import java.util.Vector;
 
 /**
- *
- * @author Mickael Gaillard <mick.gaillard@gmail.com>
+ * Utility for load JNI.
  */
 public abstract class NativeUtils {
