+++ rcljava/src/main/java/org/ros2/rcljava/exception/ImplementationAlreadyImportedException.java
@@ -1,4 +1,5 @@
-/* Copyright 2016 Open Source Robotics Foundation, Inc.
+/* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +19,6 @@ package org.ros2.rcljava.exception;
  * Raised on select_rmw_implemenation() after import_rmw_implementation() has
  * been called.
  *
- * @author Mickael Gaillard <mick.gaillard@gmail.com>
  */
 public class ImplementationAlreadyImportedException extends Exception {
