+++ rcljava/src/main/java/org/ros2/rcljava/exception/NoImplementationAvailableException.java
@@ -1,4 +1,5 @@
-/* Copyright 2016 Open Source Robotics Foundation, Inc.
+/* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +17,6 @@ package org.ros2.rcljava.exception;
 
 /**
  * Raised when there is no rmw implementation with a Java extension available.
- *
- * @author Mickael Gaillard <mickael.gaillard@gmail.com>
  *
  */
 public class NoImplementationAvailableException extends Exception {
