+++ rcljava/src/main/java/org/ros2/rcljava/node/parameter/ParameterVariant.java
@@ -1,4 +1,5 @@
-/* Copyright 2016 Open Source Robotics Foundation, Inc.
+/* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,7 +23,6 @@ import rcl_interfaces.msg.ParameterValue;
  * Parameter Variant.
  *
  * @param <T> Parameter Type.
- * @author Mickael Gaillard <mick.gaillard@gmail.com>
  */
 public class ParameterVariant<T> {
