+++ rcljava/src/main/java/org/ros2/rcljava/node/service/BiConsumer.java
@@ -1,4 +1,5 @@
 /* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.ros2.rcljava.node.service;
 
 /**
@@ -23,6 +23,7 @@ package org.ros2.rcljava.node.service;
  * @param <U> - the type of the second input to the operation
  */
 public interface BiConsumer<T, U> {
+
   /**
    * Performs this operation on the given argument.
    *
@@ -30,4 +31,5 @@ public interface BiConsumer<T, U> {
    * @param input2 - the second input argument
    */
   void accept(T input1, U input2);
+
 }
