+++ rcljava/src/main/java/org/ros2/rcljava/node/service/RMWRequestId.java
@@ -1,4 +1,5 @@
-/* Copyright 2016 Open Source Robotics Foundation, Inc.
+/* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,14 +16,11 @@
 package org.ros2.rcljava.node.service;
 
 /**
- * Service Consumer.
- *
- * @param <T> Service Request Type.
- * @param <U> Service Response Type.
- * @author Mickael Gaillard <mick.gaillard@gmail.com>
+ * Service Request ID.
  */
-public interface ServiceConsumer<T, U> {
+public class RMWRequestId {
 
-    public void call(T request, U response);
+    public byte[] writerGUID = new byte[16];
+    public long sequenceNumber;
 
 }
