+++ rcljava/src/main/java/org/ros2/rcljava/node/topic/Topics.java
@@ -1,7 +1,29 @@
+/* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.ros2.rcljava.node.topic;
 
+/**
+ * Topics built-in.
+ */
 public class Topics {
 
+    /** Standard ros2 out. */
     public static final String ROSOUT = "/rosout";
+
+    /** Parameter events topic. */
+    public static final String PARAM_EVENT = "parameter_events";
 
 }
