+++ rcljava/src/main/java/org/ros2/rcljava/qos/policies/History.java
@@ -1,4 +1,5 @@
 /* Copyright 2016 Esteve Fernandez <esteve@apache.org>
+ * Copyright 2016 Mickael Gaillard <mick.gaillard@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,9 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.ros2.rcljava.qos.policies;
 
+/**
+ * History enum.
+ *
+ */
 public enum History implements QoSPolicy {
   SYSTEM_DEFAULT(0),
   KEEP_LAST(1),
