--- rcljava/src/test/java/org/ros2/rcljava/test/NodeTest.java
@@ -1,313 +1,0 @@
-/* Copyright 2016 Open Source Robotics Foundation, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.ros2.rcljava.test;
-
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Ignore;
-import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import org.ros2.rcljava.qos.QoSProfile;
-import org.ros2.rcljava.RCLJava;
-import org.ros2.rcljava.RMWRequestId;
-import org.ros2.rcljava.node.Node;
-import org.ros2.rcljava.node.service.Client;
-import org.ros2.rcljava.node.service.Service;
-import org.ros2.rcljava.node.service.TriConsumer;
-import org.ros2.rcljava.node.topic.Consumer;
-import org.ros2.rcljava.node.topic.Publisher;
-import org.ros2.rcljava.node.topic.Subscription;
-
-import java.util.HashMap;
-
-/**
- *
- * @author Mickael Gaillard <mick.gaillard@gmail.com>
- */
-public class NodeTest {
-    private static final Logger logger = LoggerFactory.getLogger(NodeTest.class);
-
-    @BeforeClass
-    public static void setUp() {
-
-    }
-
-    @Test
-    public void testCreateNode() {
-        boolean test = true;
-        Node node = null;
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testNode");
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertEquals("Bad result", node, node);
-    }
-
-    @Test
-    public void testDestroyNode() {
-        boolean test = true;
-        Node node = null;
-
-        RCLJava.rclJavaInit();
-        node = RCLJava.createNode("testNode");
-        try {
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertEquals("Bad result", node, node);
-    }
-
-    @Test
-    public void testGetNodeName() {
-        boolean test = true;
-        Node node = null;
-        String nodeName = null;
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testNodeName");
-            nodeName = node.getName();
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertEquals("Bad result", "testNodeName", nodeName);
-    }
-
-    @Ignore
-    @Test
-    public void testPublisher() {
-        boolean test = true;
-        Node node = null;
-        Publisher<std_msgs.msg.String> pub = null;
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testPublisher");
-            pub = node.<std_msgs.msg.String>createPublisher(
-                    std_msgs.msg.String.class,
-                    "testChannel",
-                    QoSProfile.DEFAULT);
-
-            pub.dispose();
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertNotNull("Bad result", pub);
-    }
-
-    @Ignore
-    @Test
-    public void testSubscription() {
-        boolean test = true;
-        Node node = null;
-        Subscription<std_msgs.msg.String> sub = null;
-
-        Consumer<std_msgs.msg.String> callback = new Consumer<std_msgs.msg.String>() {
-            @Override
-            public void accept(std_msgs.msg.String msg) { }
-        };
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testSubscription");
-            sub = node.<std_msgs.msg.String>createSubscription(
-                    std_msgs.msg.String.class,
-                    "testChannel",
-                    callback,
-                    QoSProfile.DEFAULT);
-
-          //TODO sub.dispose();
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertNotNull("Bad result", sub);
-    }
-
-    @Ignore
-    @Test
-    public void testClient() {
-        boolean test = true;
-        Node node = null;
-        Client<std_msgs.msg.String> clt = null;
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testClient");
-            clt = node.<std_msgs.msg.String>createClient(
-                    std_msgs.msg.String.class,
-                    "testChannel",
-                    QoSProfile.DEFAULT);
-          //TODO clt.dispose();
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertNotNull("Bad result", clt);
-    }
-
-    @Ignore
-    @Test
-    public void testService() {
-        boolean test = true;
-        Node node = null;
-        Service<std_msgs.msg.String> srv = null;
-
-        TriConsumer<RMWRequestId, std_msgs.msg.String, std_msgs.msg.String> callback =
-                new TriConsumer<RMWRequestId, std_msgs.msg.String, std_msgs.msg.String>() {
-            @Override
-            public void accept(RMWRequestId header, std_msgs.msg.String request, std_msgs.msg.String response) { }
-        };
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testSubscription");
-            srv = node.<std_msgs.msg.String>createService(
-                    std_msgs.msg.String.class,
-                    "testChannel",
-                    callback,
-                    QoSProfile.DEFAULT);
-          //TODO srv.dispose();
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertNotNull("Bad result", srv);
-    }
-
-    @Test
-    public void testGraphPublisherCount() {
-        boolean test = true;
-        int count = -2;
-        Node node = null;
-        Publisher<std_msgs.msg.String> pub = null;
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testPublisher");
-
-            count = node.countPublishers("testChannel");
-            Assert.assertEquals("Bad result", 0, count);
-            // TODO
-//            pub = node.<std_msgs.msg.String>createPublisher(
-//                    std_msgs.msg.String.class,
-//                    "testChannel",
-//                    QoSProfile.PROFILE_DEFAULT);
-//            count = node.countPublishers("testChannel");
-//            Assert.assertEquals("Bad result", 1, count);
-//            pub.dispose();
-            count = node.countPublishers("testChannel");
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertEquals("Bad result", 0, count);
-    }
-
-    @Test
-    public void testGraphSubscriptionCount() {
-        boolean test = true;
-        int count = -2;
-        Node node = null;
-        Subscription<std_msgs.msg.String> sub = null;
-
-        Consumer<std_msgs.msg.String> callback = new Consumer<std_msgs.msg.String>() {
-            @Override
-            public void accept(std_msgs.msg.String msg) { }
-        };
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testSubscription");
-            count = node.countPublishers("testChannel");
-            Assert.assertEquals("Bad result", 0, count);
-            // TODO
-//            sub = node.<std_msgs.msg.String>createSubscription(
-//                    std_msgs.msg.String.class,
-//                    "testChannel",
-//                    callback,
-//                    QoSProfile.PROFILE_DEFAULT);
-//            count = node.countPublishers("testChannel");
-//            Assert.assertEquals("Bad result", 1, count);
-//            sub.dispose();
-            count = node.countPublishers("testChannel");
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertEquals("Bad result", 0, count);
-    }
-
-    @Test
-    public void testGraphGetTopics() {
-        boolean test = true;
-        Node node = null;
-        HashMap<String, String> topics = null;
-
-        RCLJava.rclJavaInit();
-        try {
-            node = RCLJava.createNode("testSubscription");
-            topics = node.getTopicNamesAndTypes();
-
-            node.dispose();
-        } catch (Exception e) {
-            test = false;
-        }
-        RCLJava.shutdown();
-
-        Assert.assertTrue("Expected Runtime error.", test);
-        Assert.assertEquals("Bad result", 0, topics.size());
-    }
-
-    //TODO Test Parameters
-
-
-}
