--- rcljava/src/test/java/org/ros2/rcljava/test/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Packages of testing rcljava
- *
- * @author Mickael Gaillard <mick.gaillard@gmail.com>
- */
-package org.ros2.rcljava.test;
